compilation fix

+++ examples/chef/common/chef-pump.cpp
@@ -123,7 +123,7 @@ DataModel::Nullable<T> LevelToSetpoint(DataModel::Nullable<uint8_t> level, DataM
     if (level.IsNull() || !level.Value() || RangeMin.IsNull() || RangeMax.IsNull())
         return DataModel::NullNullable;
 
-    return RangeMin.Value() + (RangeMax.Value() - RangeMin.Value()) * std::min(level.ValueOr(0), 200) / 200;
+    return RangeMin.Value() + (RangeMax.Value() - RangeMin.Value()) * std::min(level.ValueOr(0), (uint8_t) 200) / 200;
 }
 
 /**
